requirements.txt
anthropic>=0.40.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0.0
python-slugify>=8.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdown>=3.6
itsdangerous>=2.1.0
